+++ node/src/test/kotlin/net/corda/node/services/config/NodeConfigurationImplTest.kt
@@ -3,10 +3,8 @@ package net.corda.node.services.config
 import net.corda.core.utilities.NetworkHostAndPort
 import net.corda.testing.ALICE
 import net.corda.testing.node.MockServices.Companion.makeTestDataSourceProperties
-import net.corda.testing.node.MockServices.Companion.makeTestDatabaseProperties
 import org.assertj.core.api.Assertions.assertThatThrownBy
 import org.junit.Test
-import java.net.URL
 import java.nio.file.Paths
 import kotlin.test.assertFalse
 import kotlin.test.assertTrue
@@ -40,7 +38,7 @@ class NodeConfigurationImplTest {
             keyStorePassword = "cordacadevpass",
             trustStorePassword = "trustpass",
             dataSourceProperties = makeTestDataSourceProperties(ALICE.name.organisation),
-            database = makeTestDatabaseProperties(),
+            database = DatabaseConfig(),
             rpcUsers = emptyList(),
             verifierType = VerifierType.InMemory,
             useHTTPS = false,