+++ docs/source/changelog.rst
@@ -91,6 +91,9 @@ UNRELEASED
 
 * ``waitForAllNodesToFinish()`` method in ``DriverDSLExposedInterface`` has instead become a parameter on driver creation.
 
+* ``database.transactionIsolationLevel`` values now follow the ``java.sql.Connection`` int constants but without the
+  "TRANSACTION_" prefix, i.e. "NONE", "READ_UNCOMMITTED", etc.
+
 .. _changelog_v1:
 
 Release 1.0